Stop suggesting that kotlin does not have class methods
I came back to this because I've used the object syntax recently without
a hitch. I started to switch the example over use the companion object syntax:

```kotlin
class Piglatin {
    companion object {
        fun translateWord(word: String): String {
            return ""
        }
    }
}
```
With the tests being able to call Piglatin.translateWord("apple") like in other languages.
But then I realized this was a lot of work and a copy change was easy.